backend/agent_researcher/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
pydantic
python-dotenv
psycopg2-binary
opentelemetry-sdk
opentelemetry-instrumentation
sqlalchemy
